ctypes: move inline test cases to a dedicated testing module
Instead of keeping the test cases at the end of libtiff_ctypes.py,
place them in test_libtiff_ctypes.py. The main advantage is that they
should be called automatically during CI.
It also reduces the actual library code, which avoids unnecessarily loading
this code every time the library is used.
